requirements.txt
fastapi
uvicorn
web3
pydantic
python-multipart
ipfshttpclient
sqlalchemy
bcrypt==4.0.1
passlib==1.7.4
python-jose
python-dotenv
